[MODIFY] 어드민 영화 정보 수정, 트레일러, 스틸컷 수정 및 삭제 완성

+++ src/main/java/com/filmfellows/cinemates/app/movie/dto/MovieInfoResponse.java
@@ -16,7 +16,7 @@ public class MovieInfoResponse {
     private String title;
     private String posterUrl;
     private LocalDate releaseDate;
-    private Long runningTime;
+    private Integer runningTime;
     private String rating;
     private String synopsis;
     private String director;
